separate ProductView
